문서 레이아웃 통합
